refactor: move connector auth registry boundary
