+++ turbo/apps/api/src/signals/routes/__tests__/connectors-type-callback.test.ts
@@ -5,7 +5,7 @@ import {
   type ConnectorOAuthClientConfig,
   type OAuthConnectorType,
 } from "@vm0/connectors/connectors";
-import { getConnectorOAuthSecretMetadata } from "@vm0/connectors/oauth-providers";
+import { getConnectorOAuthSecretMetadata } from "@vm0/connectors/auth-providers";
 import { testOauthProvider } from "@vm0/connectors/oauth-providers/providers/test-oauth-provider";
 import { agentComposes } from "@vm0/db/schema/agent-compose";
 import { connectors } from "@vm0/db/schema/connector";
