+++ turbo/packages/connectors/src/__tests__/connectors.test.ts
@@ -41,7 +41,7 @@ import {
   refreshConnectorOAuthToken,
   revokeConnectorOAuthToken,
   startConnectorOAuthDeviceAuth,
-} from "../oauth-providers/provider-registry";
+} from "../auth-providers/provider-registry";
 import { GOOGLE_OAUTH_CONNECTOR_TYPES } from "../oauth-providers/google-oauth-connectors";
 import { buildGoogleAuthorizationUrl } from "../oauth-providers/providers/google-oauth";
